restructuring the project to match the required for redux router
