+++ src/App/components/BookDetail/components/BookInfo/index.js
