+++ src/App/components/BookDetail/components/Comments/components/ShowComment/ShowComment.scss
@@ -1,0 +1,28 @@
+@import "../../../../../../../stylesheets/variables.scss";
+
+.comment {
+  display: flex;
+  margin: 20px 0 20px 25px;
+}
+
+.comment-info {
+  display: flex;
+  flex-direction: column;
+  margin-left: 10px;
+}
+
+.user-name {
+  font-size: 14px;
+  font-weight: bold;
+}
+
+.comment-date {
+  font-size: 14px;
+  font-style: italic;
+  margin-top: 5px;
+}
+
+.comment-text {
+  color: $grey-3;
+  font-size: 14px;
+}
