+++ src/App/components/BookDetail/components/Suggestions/Suggestions.scss
@@ -1,0 +1,20 @@
+@import "../../../../../stylesheets/variables.scss";
+
+.suggestions-container {
+  padding: 0 25px 30px;
+  border-bottom: 2px solid $blue-1;
+}
+
+.section-title {
+  color: $green-1;
+}
+
+.book-cover {
+  width: 85px;
+  height: 125px;
+  margin-right: 55px;
+}
+
+.no-cover {
+  background: url('../../../../../assets/default_book.svg') no-repeat center $grey-2;
+}
