+++ src/App/components/BookDetail/components/Suggestions/index.js
