+++ src/App/components/Dashboard/components/Filter/Filter.scss
@@ -1,0 +1,37 @@
+@import '../../../../../stylesheets/variables.scss';
+
+.filter {
+  align-self: flex-start;
+  margin: 40px 0;
+
+  .search-form {
+    display: flex;
+  }
+
+  .filter-value {
+    background: $grey-1;
+    border: none;
+    font-family: 'Roboto', sans-serif;
+    font-style: italic;
+    font-size: 14px;
+    height: 40px;
+    padding-left: 10px;
+    width: 240px;
+  }
+
+  .search-value {
+    display: flex;
+    font-style: italic;
+    height: 100%;
+    max-height: 40px;
+  }
+
+  .search-button {
+    background-color: $green-1;
+    border: none;
+    display: flex;
+    height: 40px;
+    justify-content: center;
+    width: 40px;
+  }
+}
