--- src/screens/App/App.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-
-import Dashboard from './screens/Dashboard';
-
-class App extends Component {
-  render() {
-    return (
-      <Dashboard />
-    );
-  }
-}
-
-export default connect()(App);
