--- src/screens/App/screens/BookDetail/components/BookInfo/BookInfo.scss
@@ -1,60 +1,0 @@
-@import "../../../../../../stylesheets/variables.scss";
-
-.book-detail {
-  padding: 20px 150px;
-  margin: 40px auto;
-  max-width: 800px;
-}
-
-.book-container {
-  border-bottom: 2px solid $blue-1;
-  display: flex;
-  justify-content: center;
-  padding-bottom: 35px;
-
-  .book-cover {
-    width: 190px;
-    height: 270px;
-    margin-right: 20px;
-  }
-
-  .no-cover {
-    background: url('../../../../../../assets/default_book.svg') no-repeat center $grey-2;
-  }
-
-  .book-info {
-    display: flex;
-    flex-direction: column;
-    max-width: 540px;
-    width: 100%;
-
-    .title {
-      margin-top: 0;
-    }
-
-    .author {
-      margin-bottom: 20px;
-    }
-
-    .year {
-      margin-bottom: 10px;
-    }
-
-    .summary {
-      font-size: 14px;
-      color: $grey-3;
-      line-height: 19px;
-    }
-
-    .rent-button {
-      border: 0;
-      background-color: $blue-1;
-      color: $white;
-      font-weight: bold;
-      font-size: 20px;
-      margin-top: 20px;
-      max-width: 190px;
-      height: 40px;
-    }
-  }
-}
