--- src/screens/App/screens/BookDetail/index.js
@@ -1,35 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-
-import books from '../../../../resources/books.json';
-import BookDetailLayout from './layout';
-
-class BookDetail extends Component {
-  state = {
-    book: null
-  }
-
-  findBook = () => {
-    const bookId = parseInt(this.props.match.params.bookId, 10);
-    const bookFound = books.filter((book) => {
-      return book.id === bookId;
-    })[0];
-    this.setBook(bookFound);
-  }
-
-  setBook = (book) => {
-    this.setState({ book });
-  }
-
-  componentWillMount = () => {
-    this.findBook();
-  }
-
-  render() {
-    return (
-      <BookDetailLayout book={ this.state.book } />
-    );
-  }
-}
-
-export default connect()(BookDetail);
