--- src/screens/App/screens/Dashboard/components/Books/Books.scss
@@ -1,39 +1,0 @@
-@import '../../../../../../stylesheets/variables.scss';
-
-.dashboard {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  margin: 0 auto;
-  max-width: 725px;
-  width: 100%;
-
-  .book {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    margin: 0 20px 65px 0;
-    max-width: 130px;
-    text-decoration: none;
-    color: $black;
-    width: 100%;
-
-    .book-cover {
-      height: 180px;
-      width: 130px;
-    }
-
-    .no-cover {
-      background: url('../../../../../../assets/default_book.svg') no-repeat center $grey-2;
-    }
-
-    .book-title {
-      font-size: 14px;
-    }
-    
-    .book-author {
-      font-size: 12px;
-    }
-  }
-}
-
