+++ src/store/index.js
@@ -2,17 +2,19 @@ import { combineReducers, createStore, applyMiddleware, compose } from 'redux';
 import { connectRoutes } from 'redux-first-router';
 import createHistory from 'history/createBrowserHistory';
 
+import routesReducer from './reducers/routesReducer';
+
 const history = createHistory();
 
-const routesMap = { 
+const routesMap = {
   LOGIN: '/login',
-  DASHBOARD: '/dashboard',      // action <-> url path
-  BOOK_DETAIL: '/book/:id'  // :id is a dynamic segment
+  DASHBOARD: '/dashboard',
+  BOOK_DETAIL: '/book/:id'
 };
 
 const { reducer, middleware, enhancer } = connectRoutes(history, routesMap);
 
-const rootReducer = combineReducers({ location: reducer });
+const rootReducer = combineReducers({ location: reducer, routesReducer });
 const middlewares = applyMiddleware(middleware);
 
 export default createStore(
