+++ src/store/reducers/routesReducer.js
@@ -1,0 +1,12 @@
+
+const routesReducer = (state = {}, action) => {
+  switch(action.type) {
+    case 'DASHBOARD':
+       return 'dashboard'
+    case 'BOOK_DETAIL':
+       return 'bookDetail'
+  }
+  return state;
+};
+
+export default routesReducer;
